refactor helper file name/s; added eslint prettier plugins; removed airbnb eslint plugin;
