+++ .prettierrc.js
@@ -3,4 +3,7 @@ module.exports = {
   tabWidth: 2,
   semi: true,
   singleQuote: false,
+  printWidth: 80,
+  bracketSpacing: true,
+  arrowParens: "always",
 };
