+++ package.json
@@ -28,11 +28,13 @@
   },
   "devDependencies": {
     "eslint": "^7.28.0",
-    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.23.4",
+    "eslint-plugin-jest": "^24.5.0",
+    "eslint-plugin-prettier": "^4.0.0",
     "jest": "^27.2.4",
     "nodemon": "^2.0.7",
     "prettier": "^2.4.1",
-    "supertest": "^6.1.6"
+    "supertest": "^6.1.3"
   }
 }
